Move all options to configure the library in a dedicated file
